pyproject.toml
[project]
name = "terrabox"
version = "0.1.0"
description = "Platform backend for the Terralink SDK"
authors = [
  { name = "Terralink", email = "xiongzhitong@gmail.com" }
]
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "httpx>=0.24",
  "pydantic>=2.6",
  "pydantic-settings>=2.0",
  "SQLAlchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "argon2-cffi>=23.1.0",
  "passlib>=1.7.4",
  "python-jose>=3.3.0",
  "email-validator>=2.1",
  "python-dotenv>=1.0",
  "redis>=5.0",
  "cryptography>=41.0.0",
  "PyGithub>=1.59.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
  "pytest",
  "httpx",
  "numpy",
  "xarray",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
pythonpath = ["src"]
filterwarnings = [
    "ignore:'crypt' is deprecated and slated for removal in Python 3.13:DeprecationWarning:passlib.utils",
]
markers = [
    "auth: marks tests as authentication related",
    "api: marks tests as API related", 
    "integration: marks tests as integration tests",
    "workflow: marks tests as workflow tests",
]
